Reorganize cpu and ppu into modules
This change removes the cpu and ppu subpackages and organizes all associated classes into a module to reduce unnecessary complexity.

Having these in sub-packages created complicated references to classes E.g. purenes.cpu.cpu.CPU, which will now be purenes.cpu.CPU.

This is not a breaking change.
