--- purenes/cpu/__init__.py
@@ -1,3 +1,0 @@
-from purenes.cpu.cpu_bus import CPUBus
-from purenes.cpu.cpu import CPU
-
