--- purenes/ppu/__init__.py
@@ -1,1 +1,0 @@
-from purenes.ppu.ppu_bus import PPUBus
