+++ setup.py
@@ -2,7 +2,7 @@ import semver
 from setuptools import setup
 from setuptools import find_packages
 
-version = semver.VersionInfo.parse('0.8.0')
+version = semver.VersionInfo.parse('0.9.0')
 
 setup(name='purenes',
       version=str(version),
